Atropisomer enhanced stereo (#37)
Atropisomer bonds are now included as part of enhanced stereo chemistry s(stereo groups). This is parsed and written to mols in the mol format, MRV format and CXSmiles format.

An atropisomer is detected when one of the atoms of the atropisomer bond is labels as part of the Stereo Group in the input file/block. When creating the output to these formats, the first atom of the atropisomer bond is marked as of the stereo group.

https://www.pivotaltracker.com/story/show/184665821