Fix ctest errors (#50)
#185093951 Get all RDKit tests running

* Fixed error in pyGraphMolWrap test
(updated to use new calling parameters for the python method. The python method has to match the C++ method signiture including parameters that have defaults)

* fixed pandas test for not fail when panda is not installed
I have no idea why this started failing - I added extra checks to see if pandas was installed.

* testMolSupplier correct canonicallization - change expected value
the corrected canonicalization revealed that this test was expected an incorrect result.  I added a test to the canon test file
that shows that two different ordering of the test mol from testMolSupplier would in fact give two different "unique" results with the old method, and they give the same unique result with the new method.

* changed test for ARY to QQQ - ARY is now allowed
ARY and other groups from ketcher and MarvinJS were added, so the test for ARY now succeeds,  THis was a test for an unknow atom type, and ARY is no longer unknown.  


---------

Co-authored-by: Tad Hurst <tad.hurst@collaborativedrug.com>