Add parameter to write wedging info only for atropisomer bonds (#65)
If the CX_BOND_CFG is added to the CXSmiles output parameter list, all wedge/hash bonds are written out - this includes wedging to atropisomers and also for all chiral atoms.   We do not need or want the wedging for chiral atoms - the chirality is handled in the smiles itself with "@" and "@@" convention.  We do need the wedging for atropisomers.

So  new paramter was added to allow just the atropisomer wedging: CX_BOND_ATROPISOMER

The following commits are included:


* Add parameter to write wedging info only for atropisomer bonds

* Added parameter for Wedges only on Atropisomers

* Added to python cases to BondStereo for atropisomers

* Error after merge conflict resolution